outputs-mismatch: Fixed crd type mismatch from crds. Refactored fixtures for tests (#55)
